+++ dataservice/__init__.py
@@ -1,7 +1,7 @@
 from dataservice.service import DataService
-from dataservice.models import ABClient
+from abclient import ABClient
 from dataservice.clients import HttpXClient
 from dataservice.models import Request, Response
 from dataservice.pipeline import Pipeline
 
-__all__ = ['ABClient', 'DataService', 'HttpXClient', 'Pipeline', 'Request', 'Response']
+__all__ = ['DataService', 'HttpXClient', 'Pipeline', 'Request', 'Response']
